Adding a comma that can make a difference
This book is useful for people who don't know math, so it might be not obvious. If they read "the square root of two divided by two", it could be "the square root of (two divided by two)". This is a common trap on too many math test I've read. If you add the comma, you can read "(the square root of two), divided by two". Whitout the comma, you could make life harder for people frustrated with maths. Let's make it easier!