Move phrase to the correct paragraph
If you begin a paragraph with the phrase "The same is true of runtime errors", the reader could think this paragraph is about runtime errors, but it's not (when I saw it, I was waiting for the runtime errors mention but it never appeared). So, saying that phrase in the previous paragraph makes more sense.